[fix] SuccessLevel 계산을 Success class에 위임
