+++ app/src/main/java/com/gojol/notto/model/data/DayWithSuccessLevelAndSelect.kt
@@ -1,9 +1,7 @@
 package com.gojol.notto.model.data
 
-import com.gojol.notto.common.SuccessLevel
-
 data class DayWithSuccessLevelAndSelect(
     val day: Int,
-    val successLevel: SuccessLevel,
+    val successLevel: Success.Level,
     val isSelected: Boolean
 )
